+++ packages/client/src/features/change-password/index.ts
@@ -1,0 +1,1 @@
+export { ChangePasswordForm } from './ui'
