+++ packages/client/src/features/profile-avatar/index.ts
@@ -1,0 +1,1 @@
+export { ProfileAvatar } from './ui'
