+++ packages/client/src/widgets/profile-info/index.ts
@@ -1,0 +1,1 @@
+export { ProfileInfo } from './ui'
